+++ src/components/common/index.tsx
@@ -1,1 +1,3 @@
 export * from './chat/RenderModelIcon'
+export * from './chat/LoadingSpinner'
+export * from './chat/LoadingDots'