+++ src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './toCamel'
 export * from './db.server'
+export * from './modelName'
+export * from './uploadImage'