+++ CHANGELOG.md
@@ -90,4 +90,5 @@ This changelog provides a detailed record of the updates and improvements made t
 - Add `Ignore above messages` feature, but had some bugs. When rendering real time messages an array, the ignored above messages will not work after refreshing the page. Because the real time messages returned id type is string, haven't stored it in the database.
 - Implement the deletion of all chatMessages in the current ChatItem.
 - Add `GPT4 vision` model.
+- Integrated `cloudinary` for image upload.